requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
loguru
openai
google-generativeai
httpx 